Fixed altRoute infinite error

+++ data_structures/Graph.h
@@ -14,7 +14,7 @@
 template <class T>
 class Edge;
 
-#define INF std::numeric_limits<double>::max()
+#define INF std::numeric_limits<int>::max()
 
 /************************* Vertex  **************************/
